requirements.txt
Pillow
pyspark
tensorflow
keras
h5py
scikit-learn